# JIT: Optimize x87 FSINCOS