… FEXCore/Source/Interface/Core/Interpreter/InterpreterOps.h
… FEXCore/include/FEXCore/Core/CoreState.h
… unittests/InstructionCountCI/AVX128/FMA4.json
… unittests/InstructionCountCI/AVX128/VEX_map2.json
… unittests/InstructionCountCI/AVX128/VEX_map2_SVE128.json
… unittests/InstructionCountCI/FlagM/HotBlocks.json
… unittests/InstructionCountCI/FlagM/VEX_map1.json
… unittests/InstructionCountCI/FlagM/x87-Crysis2Max-fmodel.json
… unittests/InstructionCountCI/FlagM/x87-HalfLife.json
… unittests/InstructionCountCI/FlagM/x87-Oblivion.json
… unittests/InstructionCountCI/FlagM/x87-Psychonauts.json
… unittests/InstructionCountCI/FlagM/x87.json
… unittests/InstructionCountCI/FlagM/x87_f64.json
… unittests/InstructionCountCI/H0F38.json
… unittests/InstructionCountCI/H0F3A.json
… unittests/InstructionCountCI/PrimaryGroup.json
… unittests/InstructionCountCI/SSE42_Strings.json
… unittests/InstructionCountCI/Secondary.json
… unittests/InstructionCountCI/Secondary_OpSize.json
… unittests/InstructionCountCI/Secondary_REP.json
… unittests/InstructionCountCI/Secondary_REPNE.json
… unittests/InstructionCountCI/VEX_map1.json
… unittests/InstructionCountCI/VEX_map1_FRINTTS.json
… unittests/InstructionCountCI/VEX_map2.json
… unittests/InstructionCountCI/VEX_map3.json
… unittests/InstructionCountCI/x87.json
… unittests/InstructionCountCI/x87_f64.json