@bylaws 2025-07-04
seems we'd have to pull in a library for sincos