@Sonicadvance1 2025-07-05
Yea, we'll need to pull in some library implementation for win32 if we want sincos to go as fast as glibc. Not too concerned about it for now.